+++ openpype/hosts/aftereffects/plugins/publish/collect_render.py
@@ -6,8 +6,8 @@ import attr
 import pyblish.api
 
 from openpype.settings import get_project_settings
-from openpype.lib import abstract_collect_render
-from openpype.lib.abstract_collect_render import RenderInstance
+from openpype.pipeline import publish
+from openpype.pipeline.publish import RenderInstance
 
 from openpype.hosts.aftereffects.api import get_stub
 
@@ -25,7 +25,7 @@ class AERenderInstance(RenderInstance):
     file_name = attr.ib(default=None)
 
 
-class CollectAERender(abstract_collect_render.AbstractCollectRender):
+class CollectAERender(publish.AbstractCollectRender):
 
     order = pyblish.api.CollectorOrder + 0.405
     label = "Collect After Effects Render Layers"
